jiawen-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jiawen"
version = "0.1.0"
description = "佳文数字生命 — 宠物AI数字孪生流水线：视频 → 姿态提取 → 3D重建 → 数字资产"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "佳文数字生命团队"},
]
keywords = ["3D reconstruction", "pet", "digital twin", "pose estimation", "gaussian splatting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "trimesh>=4.0.0",
    "pygltflib>=1.16.0",
    "scipy>=1.11.0",
    "scikit-image>=0.21.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "einops>=0.7.0",
    "gradio>=4.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
colmap = [
    "pycolmap>=0.6.0",
]
viz = [
    "open3d>=0.18.0",
    "viser>=0.1.0",
]

[project.scripts]
jiawen = "video2pet.cli:app"

[tool.setuptools.packages.find]
include = ["video2pet*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
